+++ src/cppNGS/Phenotype.cpp
@@ -1,6 +1,6 @@
 #include "Phenotype.h"
 
-Phenotype::Phenotype(QByteArray accession, QByteArray name)
+Phenotype::Phenotype(const QByteArray& accession, const QByteArray& name)
 	: accession_(accession)
 	, name_(name)
 {