+++ src/tools-TEST/data_in/NGSDMaintain_in1.sql
@@ -1,5 +1,5 @@
 
-INSERT INTO `user`(`id`, `user_id`, `password`, `user_role`, `name`, `email`, `created`, `active`) VALUES (2, 'ahmustm1', '', 'user', 'Max Mustermann', '', '2016-07-05', 1);
+INSERT INTO `user`(`id`, `user_id`, `password`, `user_role`, `name`, `email`, `created`, `active`) VALUES (99, 'ahmustm1', '', 'user', 'Max Mustermann', '', '2016-07-05', 1);
 
 INSERT INTO `device` (`id`, `type`, `name`) VALUES (1, 'MiSeq', 'Neo');
 
@@ -60,3 +60,14 @@ INSERT INTO `hpo_term` (`id`, `hpo_id`, `name`, `definition`, `synonyms`) VALUES
 INSERT INTO `hpo_genes` (`hpo_term_id`, `gene`) VALUES
 (11495, 'BRCA1'),
 (11495, 'A1BG-AS');
+
+INSERT INTO `omim_gene` (`id`, `gene`, `mim`) VALUES
+(95435, 'BRCA1', '113705'),
+(99999, 'A1BG-AS', '999999');
+
+INSERT INTO `sample_disease_info` (`id`, `sample_id`, `disease_info`, `type`, `user_id`, `date`) VALUES
+(12, 1, 'HP:0000003', 'HPO term id', 99, '2018-11-21 18:26:26'),
+(13, 1, 'C34.9', 'ICD10 code', 99, '2018-11-21 18:26:26'),
+(14, 1, '60', 'tumor fraction', 99, '2018-11-21 18:26:26'),
+(15, 2, 'G11.9', 'ICD10 code', 99, '2018-11-21 18:35:57'),
+(16, 2, 'HP:0001272', 'HPO term id', 99, '2018-11-21 18:35:57');